routes fix

+++ web/app/Api/V1/routes.php
@@ -15,20 +15,6 @@ $router->group([
      * PRIVATE ACCESS
      */
     $router->group([
-        'prefix' => '',
-        'namespace' => 'Admin',
-    ], function ($router) {
-        /**
-         * User admin
-         */
-        $router->post('/subscribers', 'SubscriberController@store');
-        $router->put('/subscribers/{id}', 'SubscriberController@update');
-    });
-
-    /**
-     * PRIVATE ACCESS
-     */
-    $router->group([
         'middleware' => 'checkUser',
     ], function ($router) {
 
@@ -42,43 +28,31 @@ $router->group([
         'namespace' => 'Admin',
         'middleware' => [
             'checkUser',
-            'checkAdmin',
-        ],
-    ], function ($router) {
-        /**
-         * Subscribers
-         */
-        $router->get('/subscribers', 'SubscriberController@index');
-        $router->get('/subscribers/{id}', 'SubscriberController@show');
-        $router->post('/subscribers', 'SubscriberController@store');
-        $router->put('/subscribers/{id}', 'SubscriberController@update');
-        $router->delete('/subscribers/{id}', 'SubscriberController@destroy');
-
-    });
-
-    /**
-     * ADMIN PANEL ACCESS
-     */
-    $router->group([
-        'prefix' => 'admins',
-        'namespace' => 'Admin',
-        'middleware' => [
-            'checkUser',
             'checkAdmin'
         ]
     ], function ($router) {
+        /**
+         * Dashboard
+         */
+        $router->get('/dashboard', 'SubscribersDashboardController@index');
+
+        /**
+         * Subscribers
+         */
+        $router->get('subscribers', 'SubscriberController@index');
+        $router->post('subscribers', 'SubscriberController@store');
+        $router->get('subscribers/{id}', 'SubscriberController@show');
+        $router->put('subscribers/{id}', 'SubscriberController@update');
+        $router->delete('subscribers/{id}', 'SubscriberController@destroy');
 
         /**
          * Admins
          */
-        $router->get('/', 'AdminController@index');
-        $router->get('/{id}', 'AdminController@show');
-        $router->post('/', 'AdminController@store');
-        $router->put('/{id}', 'AdminController@update');
-        $router->delete('/{id}', 'AdminController@destroy');
-        $router->patch('/{id}', 'AdminController@updateRole');
-
-
-        $router->get('/dashboard', 'SubscribersDashboardController@index');
+        $router->get('admins', 'AdminController@index');
+        $router->post('admins', 'AdminController@store');
+        $router->get('admins/{id}', 'AdminController@show');
+        $router->put('admins/{id}', 'AdminController@update');
+        $router->delete('admins/{id}', 'AdminController@destroy');
+        $router->patch('admins/{id}', 'AdminController@updateRole');
     });
 });
